blog delete only with token

+++ osa4/controllers/blogs.js
@@ -9,18 +9,10 @@ BlogsRouter.get('/', async (request, response) => {
   response.json(blogs)
 })
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if(authorization && authorization.toLowerCase().startsWith('bearer ')) 
-    return authorization.substring(7)
-  return null
-}
-
 BlogsRouter.post('/', async (request, response, next) => {
   const body = request.body
-  const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!token || !decodedToken.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id)
     return response.status(401).json({ error: 'token missing or invalid' })
   
   const user = await User.findById(decodedToken.id)
@@ -42,8 +34,16 @@ BlogsRouter.post('/', async (request, response, next) => {
 })
 
 BlogsRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id)
+    return response.status(401).json({ error: 'token missing or invalid' })
+  const blog = await Blog.findById(request.params.id)
+  if(blog.user.toString() === decodedToken.id){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else
+    return response.status(401).json({ error: 'wrong user' })
 })
 
 module.exports = BlogsRouter
